remove stdint

+++ src/kyber_scalar/verify.c
@@ -19,19 +19,19 @@
  *
  * Description: Compare two arrays for equality in constant time.
  *
- * Arguments:   const uint8_t *a: pointer to first byte array
- *              const uint8_t *b: pointer to second byte array
- *              size_t len:       length of the byte arrays
+ * Arguments:   const u8 *a: pointer to first byte array
+ *              const u8 *b: pointer to second byte array
+ *              u64 len:       length of the byte arrays
  *
  * Returns 0 if the byte arrays are equal, 1 otherwise
  **************************************************/
-int verify(const uint8_t *a, const uint8_t *b, size_t len) {
-	size_t i;
-	uint8_t r = 0;
+int verify(const u8 *a, const u8 *b, u64 len) {
+	u64 i;
+	u8 r = 0;
 
 	for (i = 0; i < len; i++) r |= a[i] ^ b[i];
 
-	return (-(uint64_t)r) >> 63;
+	return (-(u64)r) >> 63;
 }
 
 /*************************************************
@@ -42,13 +42,13 @@ int verify(const uint8_t *a, const uint8_t *b, size_t len) {
  *              assumes two's complement representation of negative integers.
  *              Runs in constant time.
  *
- * Arguments:   uint8_t *r:       pointer to output byte array
- *              const uint8_t *x: pointer to input byte array
- *              size_t len:       Amount of bytes to be copied
- *              uint8_t b:        Condition bit; has to be in {0,1}
+ * Arguments:   u8 *r:       pointer to output byte array
+ *              const u8 *x: pointer to input byte array
+ *              u64 len:       Amount of bytes to be copied
+ *              u8 b:        Condition bit; has to be in {0,1}
  **************************************************/
-void cmov(uint8_t *r, const uint8_t *x, size_t len, uint8_t b) {
-	size_t i;
+void cmov(u8 *r, const u8 *x, u64 len, u8 b) {
+	u64 i;
 
 #if defined(__GNUC__) || defined(__clang__)
 	// Prevent the compiler from
@@ -71,11 +71,11 @@ void cmov(uint8_t *r, const uint8_t *x, size_t len, uint8_t b) {
  *              Requires b to be in {0,1};
  *              Runs in constant time.
  *
- * Arguments:   int16_t *r:       pointer to output int16_t
- *              int16_t v:        input int16_t
- *              uint8_t b:        Condition bit; has to be in {0,1}
+ * Arguments:   i16 *r:       pointer to output i16
+ *              i16 v:        input i16
+ *              u8 b:        Condition bit; has to be in {0,1}
  **************************************************/
-void cmov_int16(int16_t *r, int16_t v, uint16_t b) {
+void cmov_int16(i16 *r, i16 v, u16 b) {
 	b = -b;
 	*r ^= b & ((*r) ^ v);
 }
